+++ src/Multiplication/MultiplicationMenu.java
@@ -5,9 +5,9 @@ public class MultiplicationMenu
 {
     Scanner input = new Scanner(System.in);
 
-    public float[] mul_func() {
+    public double[] mul_func() {
         System.out.println("\nPlease enter the first number: ");
-        float[] inputs = new float[2];
+        double[] inputs = new double[2];
         if(input.hasNextFloat())
         {
             inputs[0] = input.nextFloat();
